Adaptation to TID's Node app guidelines and linting
