+++ CHANGES_NEXT_RELEASE
@@ -1,0 +1,1 @@
+- [TASK] Adaptation to TID's Node app guidelines
