+++ bin/README.md
@@ -1,0 +1,5 @@
+Excutable folder
+================
+
+Place here the node script to be linked to the Filesystem when your app is installed
+
