+++ ghpages/index.html
@@ -1,0 +1,72 @@
+<!doctype html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="chrome=1">
+    <title>github-pages-test by Daniel Moran</title>
+
+    <link rel="stylesheet" href="stylesheets/styles.css">
+    <link rel="stylesheet" href="stylesheets/pygment_trac.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
+    <!--[if lt IE 9]>
+    <script src="//html5shiv.googlecode.com/svn/trunk/html5.js"></script>
+    <![endif]-->
+</head>
+<body>
+<div class="wrapper">
+    <header>
+        <h1>github-pages-test</h1>
+        <p>The best project ever.</p>
+
+        <p class="view"><a href="https://pdihub.hi.inet/Daniel Moran/github-pages-test">View the Project on PdiHub <small>Daniel Moran/github-pages-test</small></a></p>
+
+
+    </header>
+    <section>
+        <h3>Welcome to Your Project's GitHub Pages.</h3>
+
+        <p>
+            This is the basic template for Node.js projects documentation for Telefonica Digital. This documentation was generated
+            by the tdigital-nodejs-boilerplate Grunt template. In order to update its contents, go to the site folder of your project,
+            edit the content, and then push it to the "gh-pages" branch, using the grunt task.
+        </p>
+
+<pre><code>
+    $ grunt site
+
+</code></pre>
+
+        <p>
+            The coverage and doc Grunt commands also write their results in this folder, so you can upload them by commiting the
+            changes in the site folder after a successful coverage or doc task.
+        </p>
+
+        <h3><a href="coverage/lcov-report/index.html">Test Coverage</a></h3>
+
+        <p>In this page you will find the test coverage of the project.</p>
+
+        <h3><a href="doc/index.html">JSDoc documentation</a></h3>
+
+        <p>In this page you will find the JSDoc API documentation.</p>
+
+        <h3><a href="report/index.html">Complexity report</a></h3>
+
+        <p>In this page you will find the Plato complexity report.</p>
+
+        <h3>Authors and Contributors</h3>
+
+        <p>Authored by: @javier and @dmj.</p>
+
+        <h3>Support or Contact</h3>
+
+        <p>Having trouble with Pages? Check out the documentation at <a href="http://help.github.com/pages">http://help.github.com/pages</a> or contact <a href="mailto:support@github.com">support@github.com</a> and we’ll help you sort it out.</p>
+    </section>
+    <footer>
+        <p>This project is maintained by <a href="https://pdihub.hi.inet/dmj">dmj</a></p>
+        <p><small>Hosted on GitHub Pages &mdash; Theme by <a href="https://github.com/orderedlist">orderedlist</a></small></p>
+    </footer>
+</div>
+<script src="javascripts/scale.fix.js"></script>
+
+</body>
+</html>
