+++ ghpages/params.json
@@ -1,0 +1,7 @@
+{
+    "name": "Tdaf-tokenmanager-restapi",
+    "body": "",
+    "tagline": "Rest API for the Token Manager Lib",
+    "google": "",
+    "note": "Don't delete this file! It's used internally to help with page regeneration."
+}
