+++ lib/.jshintrc
@@ -1,0 +1,18 @@
+{
+	"curly": true,
+	"eqeqeq": true,
+	"immed": true,
+	"latedef": true,
+	"newcap": true,
+	"noarg": true,
+	"noempty": true,
+	"quotmark": "single",
+	"undef": true,
+	"unused": true,
+	"trailing": true,
+	"maxparams": 4,
+	"maxdepth": 4,
+	"camelcase": true,
+	"maxlen": 120,
+	"node": true
+}
