+++ lib/README.md
@@ -1,0 +1,6 @@
+Source code
+===========
+
+Place here the source code of your app
+
+There MUST BE a file called as the project itself, and is referenced in ../index.js. This is the entry point of your app
