+++ lib/cli.js
@@ -1,0 +1,19 @@
+'use strict';
+
+
+/**
+ * This file process should arguments the terminal
+ * @return {String}
+ *   The Readed String.
+ */
+function process() {
+  var command = 'read';
+  console.log('Fake reading terminal arguments');
+  return command;
+}
+
+
+/**
+ * The exported functions and properties
+ */
+exports.process = process;
