--- package/rpm/SOURCES/usr/sth/conf/sth_default.conf
@@ -1,97 +1,0 @@
-#!/usr/bin/env bash
-
-# Short Term History configuration variables.
-# STH is configured defining the folowing variables. 
-# All variables are optional, so, if one of this variables is not defined 
-# default value is used. 
-
-# See https://github.com/telefonicaid/IoT-STH#running-the-sth-server 
-# for further information
-
-# Note: is important to export environment variables with export
-
-# The host where the STH server will be started. Default value: "localhost".
-export STH_HOST="0.0.0.0"
-
-# The port where the STH server will be listening. Default value: 8666.
-export STH_PORT="8666"
-
-# A flag indicating if the empty results should be removed from the response. Default value: "true".
-export FILTER_OUT_EMPTY="true"
-
-# Directory where temporary files will be stored, such as the ones generated when CSV files
-#   are requested. Default value: "temp".
-export TEMPORAL_DIR="temp"
-
-# The service to be used if not sent by the Orion Context Broker in the notifications. Default value: "orion".
-export DEFAULT_SERVICE="orion"
-
-# The service path to be used if not sent by the Orion Context Broker in the notifications. Default value: "/".
-export DEFAULT_SERVICE_PATH="/"
-
-# The username to use for the database connection. Default value: "".
-export DB_USERNAME=""
-
-# The password to use for the database connection. Default value: "".
-export DB_PASSWORD=""
-
-# The URI to use for the database connection. This does not include the 'mongo://' protocol part. Two or more host can be configured due to a MongoDB replica set with a comma separating each host. Default value: "localhost:27017".
-export DB_URI="localhost:27017"
-
-# The name of the MongoDB replica set. Default value: "".
-export REPLICA_SET=""
-
-# The prefix to be added to the service for the creation of the databases. More information below. Default value: "sth".
-export DB_PREFIX="sth_"
-
-# The prefix to be added to the collections in the databases. More information below. Default value: "sth".
-export COLLECTION_PREFIX="sth_"
-
-# The default MongoDB pool size of database connections. Default value: "5".
-export POOL_SIZE="5"
-
-# The write concern policy to apply when writing data to the MongoDB database. Default value: "1". Write concenr mor info: http://docs.mongodb.org/manual/core/write-concern/
-export WRITE_CONCERN="1"
-
-# Flag indicating if the raw and/or aggregated data should be persisted. Valid values are: "only-raw", "only-aggregated" and "both". Default value: "both".
-export SHOULD_STORE="both"
-
-# Flag indicating if the raw and/or aggregated data collection names should include a hash portion. This is mostly due to MongoDB's limitation regarding the number of bytes a namespace may have (currently limited to 120 bytes). In case of hashing, information about the final collection name and its correspondence to each concrete service path, entity and (if applicable) attribute is stored in a collection named COLLECTION_PREFIX + "collection_names". Default value: "false".
-export SHOULD_HASH="false"
-
-# Data from the raw and aggregated data collections will be removed if older than the value specified in seconds. In case of raw data the reference time is the one stored in the 'recvTime' property whereas in the case of the aggregated data the reference of time is the one stored in the '_id.origin' property. Set the value to 0 or remove the property entry not to apply this time-based truncation policy. Default value: "0".
-export TRUNCATION_EXPIREAFTERSECONDS="0"
-
-# The oldest raw data (according to insertion time) will be removed if the size of the raw data collection gets bigger than the value specified in bytes. Set the value to 0 or remove the property entry not to apply this truncation policy. Take into consideration than the "size" configuration parameter is mandatory in case size collection truncation is desired as required by MongoDB. Default value: "0". Notice that this configuration parameter does not affect the aggregated data collections since MongoDB does not currently support updating documents in capped collections which increase the size of the documents. Notice also that in case of the raw data, the size-based truncation policy takes precedence over the TTL one. More concretely, if "size" is set, the value of "exporeAfterSeconds" is ignored for the raw data collections since currently MongoDB does not support TTL in capped collections.
-export TRUNCATION_SIZE="0"
-
-# The oldest raw data (according to insertion time) will be removed if the number of documents in the raw data collections goes beyond the specified value. Set the value to 0 or remove the property entry not to apply this truncation policy. Default value: "0". Notice that this configuration parameter does not affect the aggregated data collections since MongoDB does not currently support updating documents in capped collections which increase the size of the documents.
-export TRUNCATION_MAX="0"
-
-# Attribute values to one or more blank spaces should be ignored and not processed either as raw data or for
-#  the aggregated computations. Default value: "true".
-export IGNORE_BLANK_SPACES="true"
-
-# The logging level of the messages. Messages with a level equal or superior
-# to this will be logged. Allowed values are: .Default value: "info".
-export LOG_LEVEL="info"
-
-# A flag indicating if the logs should be sent to the console. Default value: true.
-export LOG_TO_CONSOLE="false"
-
-# A flag indicating if the logs should be sent to a file. Default value: true.
-export LOG_TO_FILE="true"
-
-# Maximum size in bytes of the log files. If the maximum size is reached, a new log file is created incrementing a counter used as the suffix for the log file name. Default value: undefined.
-# 20MB=20*1024*1024=20971520
-export LOG_FILE_MAX_SIZE_IN_BYTES="20971520"
-
-# The path to a directory where the log file will be searched for or created if it does not exist. Default value: "./log"
-export LOG_DIR="/var/log/sth"
-
-# The name of the file where the logs will be stored. Default value: "sth_app.log".
-export LOG_FILE_NAME="sth_default.log"
-
-# The time in seconds between proof of life logging messages informing that the server is up and running normally.
-#  Default value: "60"
-export PROOF_OF_LIFE_INTERVAL="60"
