+++ rpm/SOURCES/etc/logrotate.d/logrotate-sth.conf
@@ -1,0 +1,30 @@
+# Copyright 2015 Telefónica Investigación y Desarrollo, S.A.U
+#
+# This file is part of the Short Time Historic (STH) component
+#
+# STH is free software: you can redistribute it and/or
+# modify it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the License,
+# or (at your option) any later version.
+#
+# STH is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# See the GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public
+# License along with STH.
+# If not, seehttp://www.gnu.org/licenses/.
+#
+# For those usages not covered by the GNU Affero General Public License
+# please contact with: [german.torodelvalle@telefonica.com]
+
+/var/log/sth/*.log
+{
+      daily
+      rotate 7
+      copytruncate
+      compress
+      notifempty
+      missingok
+}
