+++ test/README.md
@@ -1,0 +1,10 @@
+Test 
+======
+
+Place test directories as per-domain-based test
+
+#Documentation
+* *unit*: Unit test
+* *e2e*: End to End
+* ...
+
